Completed Ex 4.18, Ex4.19, Ex4.20, Ex4.21

+++ Chapter4/PushdownStack/Exercise/Ex4_15/stackArray.c
@@ -32,7 +32,7 @@
   * @brief private variable representing the stack head
   * 
   */
- static int N;
+ static size_t N;
  
  
  void STACKinit(size_t capacity) {
